[Security Solution] Relocate sourcerer_paths, trim dead types and removed unused variables from sourcerer folder
