+++ x-pack/solutions/security/plugins/security_solution/public/detections/components/alerts_kpis/common/hooks.ts
@@ -14,7 +14,7 @@ import type { FieldSpec } from '@kbn/data-plugin/common';
 import { i18n } from '@kbn/i18n';
 import { useAppToasts } from '../../../../common/hooks/use_app_toasts';
 import type { GlobalTimeArgs } from '../../../../common/containers/use_global_time';
-import { getScopeFromPath } from '../../../../sourcerer/containers/sourcerer_paths';
+import { getScopeFromPath } from '../../../../data_view_manager/utils/paths';
 import { getAllFieldsByName } from '../../../../common/containers/source';
 import { isLensSupportedType } from '../../../../common/utils/lens';
 import { useBrowserFields } from '../../../../data_view_manager/hooks/use_browser_fields';
