+++ x-pack/solutions/security/plugins/security_solution/public/sourcerer/store/model.ts
@@ -8,26 +8,12 @@
 import type { BrowserFields } from '@kbn/timelines-plugin/common';
 import { EMPTY_BROWSER_FIELDS } from '@kbn/timelines-plugin/common';
 import type { DataViewSpec } from '@kbn/data-views-plugin/public';
-import type { RuntimeFieldSpec, RuntimePrimitiveTypes } from '@kbn/data-views-plugin/common';
 import type { PageScope } from '../../data_view_manager/constants';
 
 /**
- * Data related to each sourcerer scope
+ * DataView from Kibana + timelines/index_fields enhanced field data
  */
-export interface SourcererScope {
-  /** Uniquely identifies a Sourcerer Scope */
-  id: PageScope;
-  /** is an update being made to the sourcerer data view */
-  loading: boolean;
-  /** selected data view id, null if it is legacy index patterns*/
-  selectedDataViewId: string | null;
-  /** selected patterns within the data view */
-  selectedPatterns: string[];
-}
-
-export type SourcererScopeById = Record<PageScope, SourcererScope>;
-
-export interface KibanaDataView {
+export interface SourcererDataView {
   /** Uniquely identifies a Kibana Data View */
   id: string;
   /**  list of active patterns that return data  */
@@ -38,17 +24,6 @@ export interface KibanaDataView {
    * comma separated string
    */
   title: string;
-}
-
-export type RunTimeMappings =
-  | Record<string, Omit<RuntimeFieldSpec, 'type'> & { type: RuntimePrimitiveTypes }>
-  | undefined;
-
-/**
- * DataView from Kibana + timelines/index_fields enhanced field data
- */
-export interface SourcererDataView extends KibanaDataView {
-  id: string;
   /**
    * @deprecated
    * determines how we can use the field in the app
@@ -66,7 +41,7 @@ export interface SourcererDataView extends KibanaDataView {
 }
 
 /**
- * Combined data from SourcererDataView and SourcererScope to create
+ * Combined data from a Kibana data view and its selected patterns to create
  * selected data view state
  */
 export interface SelectedDataView {
@@ -79,33 +54,13 @@ export interface SelectedDataView {
   indicesExist: boolean;
   /** is an update being made to the data view */
   loading: boolean;
-  /* all selected patterns from SourcererScope['selectedPatterns'] */
-  selectedPatterns: SourcererScope['selectedPatterns'];
+  /* all selected patterns */
+  selectedPatterns: string[];
   /**
    * Easier to add this additional data rather than
    * try to extend the SelectedDataView type from DataView.
    */
   sourcererDataView: DataViewSpec;
-}
-
-/**
- * sourcerer model for redux
- */
-export interface SourcererModel {
-  /** default security-solution data view */
-  defaultDataView: SourcererDataView & { id: string; error?: unknown };
-  /** default security-solution alert data view */
-  alertDataView: SourcererDataView & { id: string; error?: unknown };
-  /** default security-solution attack data view */
-  attackDataView: SourcererDataView & { id: string; error?: unknown };
-  /** all Kibana data views, including security-solution */
-  kibanaDataViews: SourcererDataView[];
-  /** security solution signals index name */
-  signalIndexName: string | null;
-  /** security solution signal index mapping state */
-  signalIndexMappingOutdated: boolean | null;
-  /** sourcerer scope data by id */
-  sourcererScopes: SourcererScopeById;
 }
 
 export type SourcererUrlState = Partial<{
@@ -114,12 +69,6 @@ export type SourcererUrlState = Partial<{
     selectedPatterns: string[];
   };
 }>;
-
-export const initSourcererScope: Omit<SourcererScope, 'id'> = {
-  loading: false,
-  selectedDataViewId: null,
-  selectedPatterns: [],
-};
 
 export const initDataView: SourcererDataView & { id: string; error?: unknown } = {
   browserFields: EMPTY_BROWSER_FIELDS,
