+++ x-pack/solutions/security/plugins/security_solution/public/timelines/components/timeline/unified_components/index.test.tsx
@@ -12,7 +12,10 @@ import { UnifiedTimeline } from '.';
 import { TimelineId, TimelineTabs } from '../../../../../common/types/timeline';
 import { useTimelineEvents } from '../../../containers';
 import { useTimelineEventsDetails } from '../../../containers/details';
-import { mockSourcererScope } from '../../../../sourcerer/containers/mocks';
+import {
+  mockBrowserFieldsWithId,
+  mockDataViewSpec,
+} from '../../../../data_view_manager/mocks/timeline_data_view';
 import {
   createSecuritySolutionStorageMock,
   mockTimelineData,
@@ -81,7 +84,7 @@ const SPECIAL_TEST_TIMEOUT = 90000;
 const localMockedTimelineData = structuredClone(mockTimelineData);
 
 const mockDataView = new DataView({
-  spec: mockSourcererScope.sourcererDataView,
+  spec: mockDataViewSpec,
   fieldFormats: fieldFormatsMock,
 });
 
@@ -90,7 +93,7 @@ const TestComponent = (
 ) => {
   const { show, ...restProps } = props;
   const testComponentDefaultProps: ComponentProps<typeof QueryTabContent> = {
-    columns: getColumnHeaders(columnsToDisplay, mockSourcererScope.browserFields),
+    columns: getColumnHeaders(columnsToDisplay, mockBrowserFieldsWithId),
     activeTab: TimelineTabs.query,
     dataView: mockDataView,
     rowRenderers: [],
