+++ x-pack/solutions/security/plugins/security_solution/public/timelines/containers/index.tsx
@@ -19,10 +19,10 @@ import type {
   TimelineEventsAllOptionsInput,
 } from '@kbn/timelines-plugin/common/api/search_strategy';
 import type { EsHitRecord } from '@kbn/discover-utils';
+import type { RunTimeMappings } from '@kbn/timelines-plugin/common/search_strategy';
 import type { ESQuery } from '../../../common/typed_json';
 
 import type { inputsModel } from '../../common/store';
-import type { RunTimeMappings } from '../../sourcerer/store/model';
 import { useKibana } from '../../common/lib/kibana';
 import { createFilter } from '../../common/containers/helpers';
 import { timelineActions } from '../store';
